Define move generation functions that outputs moves in output paramter.
The motivation for this is to allow AllMoves to add all moves the the
same vector without having to merge vectors from different functions,
which will affect performance. If there is a performance gain from
computing the moves in parallel, then it will make more sense to have
each function return a vector a moves, but we will investigate this
later.