Make BitBoard a proper class.
The idea is to have better type-safety by avoiding problems with
different fields being the same type but serving different functions.
For example, the index of a bit, i.e. the square, is also represented as
a uint, and this can lead to errors using the incorrect variables
because they are of the same type.

Making BitBoard into a class also makes it easier to bake in common bit
manipulation into the class, e.g. such as getting the count of bits set
or the index of the first bit set.