Add static member for bishop and rook magics.
The plan is to move the responsibility of move generation to the
BoardClass, because move generation is tightly coupled to the state of
the board, and it will simplify some things. But we avoid using
non-static members, otherwise every instance of board would have the same
pointer or reference. This will require that we register the magics with
the Board class before using an instance of Board to generate moves.