Begin to implement Board::update(Move).
The function will be used to apply a move to a Board. Defined some getter
functions to get the mutable piece sets, and left a bunch of TODOs in
update() because there's a lot more that needs to be done for it to
work. For not, left it as is for things to compile.