Implement Board::quick_update().
Board::quick_update() updates the board state without updating the game
state. To compute the game state, we need to look ahead, i.e. we need to
make moves and see if the king is still under attack after we make our
move, but we want to avoid recurusion since we only need to look ahead
one move.