Implement BoardPath and incorporate it as input for MTCS and Evaluator.
In addition to the current position, the evaluator considers previous
positions leading up to the current position, and they are used as
inputs for the neural network. However, the positions are in reverse
order, with current position at positioned before previous positions.
Furthermore, during search we use a tree data structure. Hence, the
point of BoardPath is to make it easier to gather the necessary boards
in the correct order for evaluation.